fix(rust, python): time.cast(str) as strftime
